Revert "separating core from the updater"
